Move page templates from site-server into site-content package
