+++ packages/site-content/.gitignore
@@ -1,2 +1,3 @@
 /_site/
 /_dev/
+/templates/lib/
